Move viewing user uploads onto the user page sidebar
